+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 * Add `SuiteEvent` to the JSON reporter, which reports data about the suites in
   which tests are run.
 
+* Add `AllSuitesEvent` to the JSON reporter, which reports the total number of
+  suites that will be run.
+
+* Add `GroupEvent.testCount` to the JSON reporter, which reports the total
+  number of tests in each group.
+
 ## 0.12.8
 
 * Organize the `--help` output into sections.